+++ src/main/java/com/example/api_spring/repositories/LocalRepository.java
@@ -3,5 +3,5 @@ package com.example.api_spring.repositories;
 import com.example.api_spring.models.Local;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface LocalRepository extends JpaRepository<Local,Integer> {
+public interface LocalRepository extends JpaRepository<Local,Long> {
 }
